Clear All Files (except sf bundle related)
