--- Exception/IdealException.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Lens\Bundle\IdealBundle\Exception;
-
-use RuntimeException;
-
-class IdealException extends RuntimeException
-{
-}
