+++ composer.json
@@ -1,25 +1,26 @@
 {
     "name": "lensmedia/symfony-ideal",
-    "description": "'Private' bundle for handling iDeal transactions",
-    "version": "1.1.5",
-    "type": "symfony-bundle",
+    "description": "\"Private\" bundle for handling iDeal transactions",
     "license": "proprietary",
+    "type": "symfony-bundle",
+    "version": "1.1.5",
     "require": {
-        "php": ">=8",
-        "ext-dom": "*",
-        "ext-simplexml": "*",
+        "php": ">=8.1",
         "ext-openssl": "*",
-        "robrichards/xmlseclibs": "*",
-        "symfony/cache": ">=5",
-        "symfony/framework-bundle": ">=5",
-        "symfony/http-client": ">=5",
-        "symfony/yaml": ">=5",
-        "symfony/form": ">=5",
-        "symfony/validator": ">=5"
+        "symfony/framework-bundle": "^7.0",
+        "symfony/monolog-bundle": "^3.10"
+    },
+    "require-dev": {
+        "ergebnis/composer-normalize": "^2.40"
     },
     "autoload": {
         "psr-4": {
-            "Lens\\Bundle\\IdealBundle\\": ""
+            "Lens\\Bundle\\IdealBundle\\": "src/"
+        }
+    },
+    "config": {
+        "allow-plugins": {
+            "ergebnis/composer-normalize": true
         }
     }
 }
