+++ src/DependencyInjection/Configuration.php
@@ -1,0 +1,20 @@
+<?php
+
+namespace Lens\Bundle\IdealBundle\DependencyInjection;
+
+use Symfony\Component\Config\Definition\Builder\TreeBuilder;
+use Symfony\Component\Config\Definition\ConfigurationInterface;
+
+class Configuration implements ConfigurationInterface
+{
+    public function __construct(
+    ) {
+    }
+
+    public function getConfigTreeBuilder(): TreeBuilder
+    {
+        $treeBuilder = new TreeBuilder('lens_ideal');
+
+        return $treeBuilder;
+    }
+}
