+++ src/Ideal.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace Lens\Bundle\IdealBundle;
+
+final class Ideal
+{
+
+}
